pyproject.toml
[project]
name = "final-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "pandas>=2.3.3",
    "pip>=25.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
    "yfinance>=1.0",
]
